+++ src/main/java/edu/infsci2560/repositories/enemyFleetRepository.java
@@ -1,0 +1,7 @@
+package edu.infsci2560.repositories;
+import edu.infsci2560.models.enemyFleet;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+
+
+public interface enemyFleetRepository extends PagingAndSortingRepository<enemyFleet, Long> {}