+++ src/main/java/edu/infsci2560/repositories/mapdataRepository.java
@@ -1,0 +1,7 @@
+package edu.infsci2560.repositories;
+import edu.infsci2560.models.mapdata;
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+
+
+public interface mapdataRepository extends PagingAndSortingRepository<mapdata, Long> {}