+++ src/main/java/edu/infsci2560/services/enemyFleetService.java
@@ -5,9 +5,8 @@
  */
 package edu.infsci2560.services;
 
-import edu.infsci2560.models.Kancolle;
-import edu.infsci2560.models.Kancolle.CommentType;
-import edu.infsci2560.repositories.KancolleRepository;
+import edu.infsci2560.models.enemyFleet;
+import edu.infsci2560.repositories.enemyFleetRepository;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -28,14 +27,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
  * @author kolobj
  */
 @RestController
-@RequestMapping("kancolle/apijson")
-public class KancolleService {
+@RequestMapping("enemyFleet/apijson")
+public class enemyFleetService {
 
     @Autowired
-    private KancolleRepository repository;
+    private enemyFleetRepository repository;
 
     @RequestMapping(method = RequestMethod.GET, produces = "application/json")
-    public ResponseEntity<Iterable<Kancolle>> list() {
+    public ResponseEntity<Iterable<enemyFleet>> list() {
         HttpHeaders headers = new HttpHeaders();
         return new ResponseEntity<>(repository.findAll(), headers, HttpStatus.OK);
     }
@@ -47,8 +46,8 @@ public class KancolleService {
     //}
 
     @RequestMapping(method = RequestMethod.POST, consumes="application/json", produces = "application/json")
-    public ResponseEntity<Kancolle> create(@RequestBody Kancolle kan) {
+    public ResponseEntity<enemyFleet> create(@RequestBody enemyFleet ef) {
         HttpHeaders headers = new HttpHeaders();
-        return new ResponseEntity<>(repository.save(kan), headers, HttpStatus.OK);
+        return new ResponseEntity<>(repository.save(ef), headers, HttpStatus.OK);
     }
 }