+++ src/main/java/edu/infsci2560/services/enemyService.java
@@ -5,9 +5,9 @@
  */
 package edu.infsci2560.services;
 
-import edu.infsci2560.models.Kancolle;
-import edu.infsci2560.models.Kancolle.CommentType;
-import edu.infsci2560.repositories.KancolleRepository;
+import edu.infsci2560.models.enemy;
+import edu.infsci2560.models.enemy.ShipType;
+import edu.infsci2560.repositories.enemyRepository;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -28,14 +28,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
  * @author kolobj
  */
 @RestController
-@RequestMapping("kancolle/apijson")
-public class KancolleService {
+@RequestMapping("enemy/apijson")
+public class enemyService {
 
     @Autowired
-    private KancolleRepository repository;
+    private enemyRepository repository;
 
     @RequestMapping(method = RequestMethod.GET, produces = "application/json")
-    public ResponseEntity<Iterable<Kancolle>> list() {
+    public ResponseEntity<Iterable<enemy>> list() {
         HttpHeaders headers = new HttpHeaders();
         return new ResponseEntity<>(repository.findAll(), headers, HttpStatus.OK);
     }
@@ -47,8 +47,8 @@ public class KancolleService {
     //}
 
     @RequestMapping(method = RequestMethod.POST, consumes="application/json", produces = "application/json")
-    public ResponseEntity<Kancolle> create(@RequestBody Kancolle kan) {
+    public ResponseEntity<enemy> create(@RequestBody enemy e) {
         HttpHeaders headers = new HttpHeaders();
-        return new ResponseEntity<>(repository.save(kan), headers, HttpStatus.OK);
+        return new ResponseEntity<>(repository.save(e), headers, HttpStatus.OK);
     }
 }