+++ src/main/java/edu/infsci2560/services/mapService.java
@@ -1,13 +1,8 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package edu.infsci2560.services;
 
-import edu.infsci2560.models.Kancolle;
-import edu.infsci2560.models.Kancolle.CommentType;
-import edu.infsci2560.repositories.KancolleRepository;
+import edu.infsci2560.models.mapdata;
+import edu.infsci2560.repositories.mapdataRepository;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -28,14 +23,14 @@ import org.springframework.web.bind.annotation.ResponseBody;
  * @author kolobj
  */
 @RestController
-@RequestMapping("kancolle/apijson")
-public class KancolleService {
+@RequestMapping("mapdata/apijson")
+public class mapService {
 
     @Autowired
-    private KancolleRepository repository;
+    private mapdataRepository repository;
 
     @RequestMapping(method = RequestMethod.GET, produces = "application/json")
-    public ResponseEntity<Iterable<Kancolle>> list() {
+    public ResponseEntity<Iterable<mapdata>> list() {
         HttpHeaders headers = new HttpHeaders();
         return new ResponseEntity<>(repository.findAll(), headers, HttpStatus.OK);
     }
@@ -47,8 +42,8 @@ public class KancolleService {
     //}
 
     @RequestMapping(method = RequestMethod.POST, consumes="application/json", produces = "application/json")
-    public ResponseEntity<Kancolle> create(@RequestBody Kancolle kan) {
+    public ResponseEntity<mapdata> create(@RequestBody mapdata map) {
         HttpHeaders headers = new HttpHeaders();
-        return new ResponseEntity<>(repository.save(kan), headers, HttpStatus.OK);
+        return new ResponseEntity<>(repository.save(map), headers, HttpStatus.OK);
     }
 }